[Chore] combine `io.ts` into `stream.ts`
